ecosky-back/requirements.txt
aiosignal==1.3.2
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
dm-tree==0.1.8
Farama-Notifications==0.0.4
filelock==3.17.0
Flask==2.2.3
Flask-Cors==3.0.10
frozenlist==1.5.0
fsspec==2024.12.0
gymnasium==0.28.1
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
jax-jumpy==1.0.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
networkx==3.4.2
numpy==1.24.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pyarrow==19.0.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
pytz==2024.2
PyYAML==6.0.2
ray[rllib]==2.9.3 
referencing==0.36.2
requests==2.28.2
rich==13.9.4
rpds-py==0.22.3
scikit-image==0.25.0
scipy==1.15.1
shellingham==1.5.4
six==1.17.0
tenacity==8.2.2
tensorboardX==2.6.2.2
tifffile==2025.1.10
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
Werkzeug==3.1.3
